+++ modules/prelude-erlang.el
@@ -38,7 +38,8 @@
   :type 'string
   :safe 'stringp)
 
-(require 'projectile)
+(when prelude-projectile
+  (require 'projectile))
 
 (when (require 'erlang-start nil t)
 
@@ -49,8 +50,9 @@
     (add-to-list 'load-path wrangler-path)
     (require 'wrangler)))
 
-(add-hook 'erlang-mode-hook (lambda ()
-                              (setq erlang-compile-function 'projectile-compile-project)))
+(when prelude-projectile
+  (add-hook 'erlang-mode-hook (lambda ()
+                                (setq erlang-compile-function 'projectile-compile-project))))
 
 (provide 'prelude-erlang)