+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add `prelude-vertico` module. Vertico a simpler alternative to `ivy-mode` and supersedes Selectrum.
 - [#1421](https://github.com/bbatsov/prelude/issues/1421): Make it possible to configure the TypeScript format action using `prelude-ts-format-action`.
 - [#1354](https://github.com/bbatsov/prelude/issues/1354): Remove default `C--` and `C-+` keybindings to increase/decrease the font size.
+- Add `prelude-projectile` user option, allowing Projectile integration to be disabled.
 
 ### Changes